requirements.txt
# all my requirement used
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.2